+++ index.html
@@ -97,7 +97,6 @@
     </div>
   </footer>
 
-
   <script src="./index.js" type="module"></script>
 </body>
 </html>
